+++ js/app.js
@@ -1,18 +1,22 @@
-import { appEventListener } from "./appEventListener.js";
-import { highlightColor, guiManager, createGUI } from "./gui.js";
-import GSDB from "./gsdb.js";
 import EventBus from "./eventBus.js";
+import GSDB from "./gsdb.js";
 import EnsembleManager from "./ensembleManager.js";
 import ColorMapManager from "./colorMapManager.js";
 import Parser from "./parser.js";
 import SceneManager, {sceneManagerConfigurator} from "./sceneManager.js";
 import DataValueMaterialProvider from "./dataValueMaterialProvider.js";
-import {appleCrayonColorRGB255, appleCrayonColorThreeJS} from "./color.js";
 import PointCloud from "./pointCloud.js";
 import Noodle from "./noodle.js";
 import BallAndStick from "./ballAndStick.js";
+import GUIManager from "./guiManager.js";
+import { appEventListener } from "./appEventListener.js";
+import { highlightColor, createGUI } from "./gui.js";
+import { appleCrayonColorRGB255, appleCrayonColorThreeJS } from "./color.js";
+
+let gsdb;
 
 let eventBus = new EventBus();
+
 let pointCloud;
 let noodle;
 let ballAndStick;
@@ -21,13 +25,15 @@ let colorMapManager;
 let parser;
 let sceneManager;
 let dataValueMaterialProvider;
-let gsdb;
 let appWindowWidth;
 let appWindowHeight;
+let guiManager;
 
 document.addEventListener("DOMContentLoaded", event => {
 
-    const container = document.getElementById('spacewalk_canvas_container');
+    gsdb = new GSDB(parser);
+
+    parser = new Parser();
 
     pointCloud = new PointCloud();
 
@@ -40,17 +46,17 @@ document.addEventListener("DOMContentLoaded", event => {
     colorMapManager = new ColorMapManager();
     colorMapManager.configure();
 
-    parser = new Parser();
-
-    sceneManager = new SceneManager(sceneManagerConfigurator({ container, highlightColor }));
-
     dataValueMaterialProvider = new DataValueMaterialProvider({ width: 2048, height: 64, colorMinimum: appleCrayonColorRGB255('silver'), colorMaximum: appleCrayonColorRGB255('blueberry'), highlightColor:appleCrayonColorThreeJS('maraschino')  });
 
-    gsdb = new GSDB(parser);
+    const container = document.getElementById('spacewalk_canvas_container');
+    sceneManager = new SceneManager(sceneManagerConfigurator({ container, highlightColor }));
+
+    guiManager = new GUIManager({ $button: $('#spacewalk_ui_manager_button'), $panel: $('#spacewalk_ui_manager_panel') });
+
+    sceneManager.setRenderStyle( guiManager.getRenderStyle() );
 
     createGUI(container);
 
-    sceneManager.setRenderStyle( guiManager.getRenderStyle() );
 
     $(window).on('resize.app', () => {
         let { width, height } = container.getBoundingClientRect();
@@ -74,4 +80,4 @@ let renderLoop = () => {
     sceneManager.render();
 };
 
-export { appWindowWidth, appWindowHeight, eventBus, pointCloud, noodle, ballAndStick, ensembleManager, colorMapManager, parser, sceneManager, dataValueMaterialProvider };
+export { appWindowWidth, appWindowHeight, eventBus, pointCloud, noodle, ballAndStick, ensembleManager, colorMapManager, parser, sceneManager, dataValueMaterialProvider, guiManager };
