+++ js/colorRampPanel.js
@@ -1,8 +1,7 @@
 import ColorRampMaterialProvider from "./colorRampMaterialProvider.js";
 import { setMaterialProvider } from './utils.js';
-import { guiManager } from './gui.js';
 import Panel from './panel.js';
-import { eventBus } from "./app.js";
+import { eventBus, guiManager } from "./app.js";
 
 class ColorRampPanel extends Panel {
