+++ js/contactFrequencyMapPanel.js
@@ -1,8 +1,8 @@
 import KDBush from '../node_modules/kd3d/js/index.js'
 import { clamp } from "./math.js";
-import { hideSpinner, showSpinner, guiManager } from './gui.js';
+import { hideSpinner, showSpinner } from './gui.js';
 import Panel from "./panel.js";
-import { colorMapManager, ensembleManager } from "./app.js";
+import { guiManager, colorMapManager, ensembleManager } from "./app.js";
 import {threeJSColorToRGB255} from "./color";
 import { drawWithSharedUint8ClampedArray } from "./utils.js";
 import EnsembleManager from "./ensembleManager.js";
