+++ js/gnomon.js
@@ -1,8 +1,7 @@
 import * as THREE from "../node_modules/three/build/three.module.js";
-import { guiManager } from "./gui.js";
 import { appleCrayonColorThreeJS, appleCrayonColorRGB255, rgb255String, threeJSColorToRGB255 } from "./color.js";
 import { numberFormatter } from "./utils.js";
-import { eventBus } from "./app.js";
+import { eventBus, guiManager } from "./app.js";
 
 class Gnomon extends THREE.AxesHelper {
