+++ js/igv/IGVPanel.js
@@ -13,8 +13,6 @@ class IGVPanel extends Panel {
 
     constructor ({ container, panel, isHidden }) {
 
-        // const isHidden = guiManager.isPanelHidden($(panel).attr('id'));
-
         const xFunction = (cw, w) => {
             return (cw - w)/2;
         };
