+++ js/juicebox/juiceboxPanel.js
@@ -6,8 +6,6 @@ class JuiceboxPanel extends Panel {
 
     constructor ({ container, panel, isHidden }) {
 
-        // const isHidden = guiManager.isPanelHidden($(panel).attr('id'));
-
         const xFunction = (cw, w) => {
             return (cw - w)/2;
         };
